make RiddleService an interface
