+++ RiddleBot/src/main/java/io/kemper/service/RiddleService.java
@@ -2,30 +2,10 @@ package io.kemper.service;
 
 import io.kemper.domain.Riddle;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Random;
+/** Interface for getting riddles. Should be implemented with a Singleton **/
+public interface RiddleService {
 
-public class RiddleService {
+    public Riddle getRandomRiddle();
 
-    List<Riddle> riddles;
-    Random randy;
-
-    public RiddleService() {
-        this.randy = new Random();
-
-        this.riddles = new ArrayList<>();
-        this.riddles.add(new Riddle(0, "what is AWS lambda?", "AWSome!!!"));
-        this.riddles.add(new Riddle(1, "Riddle 1", "Answer 1"));
-        this.riddles.add(new Riddle(2, "Riddle 2", "Answer 2"));
-    }
-
-    public Riddle getRandomRiddle() {
-        int id = randy.nextInt(this.riddles.size());
-        return riddles.get(id);
-    }
-
-    public Riddle getRiddle(Integer id) {
-        return riddles.get(id);
-    }
+    public Riddle getRiddle(Integer id);
 }
