+++ RiddleBot/src/main/java/io/kemper/service/SimpleRiddleService.java
@@ -1,0 +1,31 @@
+package io.kemper.service;
+
+import io.kemper.domain.Riddle;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Random;
+
+public class SimpleRiddleService implements RiddleService {
+
+    List<Riddle> riddles;
+    Random randy;
+
+    public SimpleRiddleService() {
+        this.randy = new Random();
+
+        this.riddles = new ArrayList<>();
+        this.riddles.add(new Riddle(0, "what is AWS lambda?", "AWSome!!!"));
+        this.riddles.add(new Riddle(1, "Riddle 1", "Answer 1"));
+        this.riddles.add(new Riddle(2, "Riddle 2", "Answer 2"));
+    }
+
+    public Riddle getRandomRiddle() {
+        int id = randy.nextInt(this.riddles.size());
+        return riddles.get(id);
+    }
+
+    public Riddle getRiddle(Integer id) {
+        return riddles.get(id);
+    }
+}
